Install zkllvm-libc.ll and related headers by default
Currently, zkllvm-libc.ll and its headers are installed as a separate
package, which has to be manually built.

When doing "cmake && make && make install" in the zkllvm root, I'm
seemingly getting everything except the libc. It's not very
convenient, as people who build manually kind of expect this approach
to just work.

I've added a default installation target that does the right thing in
this regard.